% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{header_section}
\alias{header_section}
\title{Add a tab panel header}
\usage{
header_section(title, description)
}
\arguments{
\item{title}{Section title}

\item{description}{Sesction description}
}
\description{
This function adds a header section to a Mayo shiny app.  It is only meant
    to be used with the the theme \code{horizon}
}
