% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{main_header}
\alias{main_header}
\title{Add a main header}
\usage{
main_header(logo_text)
}
\arguments{
\item{logo_text}{Text to appear in the main header logo}
}
\description{
This function adds a primary header to a Mayo shiny app.  It is only meant
    to be used with the the theme \code{modern}
}
