% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayo_loader.R
\name{mayo_loader}
\alias{mayo_loader}
\title{A custom Mayo themed loading gif}
\usage{
mayo_loader(ui_element)
}
\arguments{
\item{ui_element}{A UI element that should include a loader while the
corresponding output is being calculated}
}
\description{
Add a loader that shows when an output element is calculating
}
