% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_mayoshiny.R
\name{mayo_shiny}
\alias{mayo_shiny}
\alias{use_mayoshiny}
\title{Shiny Template for Mayo}
\usage{
mayo_shiny(path, theme = c("classic", "horizon"), ...)

use_mayoshiny(path, theme = c("classic", "horizon"), ...)
}
\arguments{
\item{path}{A directory path for where to initialize the Shiny app.}

\item{theme}{Name of a theme.}

\item{...}{Other arguments (not in use at this time)}
}
\description{
An easy-to-use, accessible template with Mayo colors and logo.
}
\details{
Normally called when creating a new project in R Studio.
   \code{mayoshiny} has been replaced by \code{use_mayoshiny}
   and will be removed in future versions.
}
