% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayoshiny.R
\docType{package}
\name{mayoshiny}
\alias{mayoshiny}
\alias{mayoshiny-package}
\title{Templates for Shiny Apps at Mayo Clinic}
\description{
Templates for Shiny Apps at Mayo Clinic.
}
\section{Functions}{


\code{\link{mayo_shiny}}

\code{\link{mayoshinytheme}}
}

\examples{
library(mayoshiny)

}
