% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayoshiny_example.R
\name{mayoshiny_example}
\alias{mayoshiny_example}
\title{View example of a Mayo Shiny template}
\usage{
mayoshiny_example(theme = c("classic", "horizon"))
}
\arguments{
\item{theme}{Name of a theme.}
}
\description{
Run a local instance of a Shiny template
}
