% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayoshinytheme.R
\name{mayoshinytheme}
\alias{mayoshinytheme}
\title{Use a Mayo Shiny Theme}
\usage{
mayoshinytheme()
}
\value{
Returns a file path reference to the Mayo Shiny CSS file, and attaches that path so shiny knows about it.
}
\description{
Style a Shiny app with Mayo colors.
}
\details{
Use this as the \code{theme = } argument to the main shiny page-builder functions, e.g., \code{\link[shiny]{fluidPage}}.
}
\examples{
\dontrun{
shinyApp(
  ui = fluidPage(theme = mayoshinytheme(), ...),
  server = function(input, output) { }
)
}
}
\seealso{
\code{\link[shinythemes]{shinytheme}}
}
