% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayo_shiny.R
\name{mayo_shiny}
\alias{mayo_shiny}
\title{Shiny Template for Mayo}
\usage{
mayo_shiny(path, ...)
}
\arguments{
\item{path}{A directory path for where to initialize the Shiny app.}

\item{...}{Other arguments (not in use at this time)}
}
\description{
An easy-to-use, accessible template with Mayo colors and logo.
}
\details{
This function is normally called when creating a new project in R Studio.
}
